[plugin-web-app] Align JS executing steps wording
